Improved pagination

+++ crates/core/src/bookmark/bookmark_repository.rs
@@ -41,7 +41,7 @@ pub struct BookmarkParams {
     pub tags: Option<Vec<Uuid>>,
     pub user_id: Option<Uuid>,
     pub cursor: Option<DateTime<Utc>>,
-    pub limit: Option<u64>,
+    pub limit: Option<usize>,
     pub with_tags: bool,
 }
